requirements.txt
matplotlib==3.7.3
opencv-python==4.8.1.78
openpyxl==3.1.2
pandas==2.0.3
pyyaml==6.0.1
scikit-learn==1.3.1
scipy==1.10.1
tqdm==4.66.1
torchsummary==1.5.1
wandb==0.16.4
numba==0.58.1
tifffile==2023.7.10
fastremap==1.14.1
natsort==8.4.0
notebook==6.5.4